+++ metrics/src/main/java/algorithms/MedianDivergenceComputer.java
@@ -91,7 +91,7 @@ public final class MedianDivergenceComputer {
 			Point2D p1 = ptIt.next();
 			Point2D p2 = ptIt.next();
 	
-			float dist = (float) p1.distance(p2);
+			double dist = p1.distance(p2);
 			
 			Color colorA = colormap.getColor(p1.getX(), p1.getY());
 			Color colorB = colormap.getColor(p2.getX(), p2.getY());
@@ -100,7 +100,7 @@ public final class MedianDivergenceComputer {
 			double cdist = colorDiff(colorA, colorB);
 			
 			// filter zero divisions, as long as the value distance is small
-			// DON'T protect colormaps that contain double colors
+			// DON'T protect colormaps that contain duplicate colors
 			if (cdist == 0 && dist < 0.05)
 				continue;
 			double ratio = cdist / dist;
